Create main.py
This Python script generates a random password based on user-specified criteria. Users can choose the number of letters, symbols, and numbers they want in their password.

The script provides two options:

Easy Level: In this level, the order of characters in the generated password is not randomized.
Hard Level: In this level, the order of characters in the generated password is randomized.